cobalt/kernel: pipeline: add abstract synthetic IRQ API
Add wrappers to create "synthetic IRQs" the I-pipe way (used to be
called "virtual IRQs" there). Those interrupt channels can only be
trigged by software, in order to run a handler on the proper execution
stage (in-band or out-band).

We reuse the Dovetail naming convention in this code: in-band meaning
"secondary mode", out-of-band meaning "primary mode" in the old
jargon.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>